pyproject.toml
[tool.poetry]
name = "olympic-paris"
version = "0.1.0"
description = ""
authors = ["wijez <viet.info.43@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
httpx = "^0.27.0"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.4.0"
asyncio = "^3.4.3"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
pyjwt = "^2.9.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
fastapi-mail = "^1.4.1"
redis = "^5.0.8"
celery = "^5.4.0"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
